+++ src/errorHandlers/index.ts
@@ -1,3 +1,4 @@
 export * from './handle400WithValidationsFunc';
 export * from './handle404Func';
 export * from './handle4xxFunc';
+export * from './handleHttpClientErrors';