+++ src/index.ts
@@ -5,6 +5,8 @@ export * from './extenders';
 export * from './httpInterceptors';
 export * from './interfaces';
 export * from './misc/classes/clientValidationError';
+export * from './misc/classes/httpClientError';
+export * from './misc/classes/httpNotFoundError';
 export * from './misc/classes/internalServerErrorInfo';
 export * from './misc/classes/restClientError';
 export * from './misc/classes/restNotFoundError';