+++ src/interfaces/index.ts
@@ -4,6 +4,7 @@ export * from './errorWithStack/errorWithStack.interface';
 export * from './errorWithUrl/errorWithUrl.interface';
 export * from './handle4xxOptions/handle4xxOptions.interface';
 export * from './httpClientError/httpClientError.interface';
+export * from './httpClientErrorOptions/httpClientErrorOptions.interface';
 export * from './httpClientErrors/httpClientErrors.interface';
 export * from './httpClientPropertyValidationError/httpClientPropertyValidationError.interface';
 export * from './httpClientValidationErrors/httpClientValidationErrors.interface';