+++ src/misc/types.ts
@@ -1,7 +1,7 @@
 import {HttpErrorResponse} from '@angular/common/http';
 import {Func1, Func4Rest} from '@jscrpt/common';
 
-import {Handle4xxOptions, HttpClientError, HttpClientValidationErrors} from '../interfaces';
+import {Handle4xxOptions, XXXHttpClientError, HttpClientValidationErrors} from '../interfaces';
 
 /**
  * Response mapper for http client errors that are converted to array of error messages
@@ -22,5 +22,5 @@ export type HttpClientErrorCustomHandler<TError = unknown,
                                                                                             HttpErrorResponse,
                                                                                             Handle4xxOptions,
                                                                                             Func1<TError, HttpErrorResponse>,
-                                                                                            Func1<TClientError, HttpClientError>,
-                                                                                            [Func1<TClientValidationError, HttpClientError>?]>;
+                                                                                            Func1<TClientError, XXXHttpClientError>,
+                                                                                            [Func1<TClientValidationError, XXXHttpClientError>?]>;