+++ src/rxjsOperators/index.ts
@@ -1,3 +1,4 @@
+export * from './catchHttpClientError.operator';
 export * from './handle400WithValidations.operator';
 export * from './handle404.operator';
 export * from './handle4xx.operator';